requirements.txt
pdf2image
pdfminer.six
Pillow
PyPDF2
tqdm